refactor: updates vote, schema and user controllers(pagination, auth middleware)

+++ src/modules/stream/streamControllers.ts
@@ -3,6 +3,8 @@
 import { CreateStreamInput, GetStreamParams , UpdateStreamInput, GetStreamsQuery } from "./streamSchema";
 import prisma from "../../utils/prisma";
 import { FastifyReply, FastifyRequest } from "fastify";
+import { PaginationQuery } from "../../utils/globalSchemas";
+import { streamExists } from "../../utils/stream";
 
 // 
 
@@ -13,12 +15,12 @@ export async function createStream(request: FastifyRequest<{ Body: CreateStreamI
             data: {
                 thumbnailUrl: input.thumbnailUrl,
                 startTimestamp: new Date(input.startTimestamp),
-                endTimestamp: new Date(input.endTimestamp),
-                storageUrl: input.storageUrl || "", 
+                endTimestamp: new Date(input.endTimestamp?input.endTimestamp:input.startTimestamp), // If endTimestamp is not provided, set it to the current date (now)
+                storageUrl: input.storageUrl, 
                 title: input.title,
                 description: input.description,
-                tags: input.tags || "",
-                userUserId: input.userUserId 
+                tags: input.tags,
+                userUserId: request.user.userId 
             }
         });
         return reply.status(200).send(stream);
@@ -29,16 +31,17 @@ export async function createStream(request: FastifyRequest<{ Body: CreateStreamI
 }
 
 
-export async function getMyStream(request: FastifyRequest<{Params:{userUserId:string}}>, reply: FastifyReply) {
+export async function getMyStream(request: FastifyRequest<{Querystring:PaginationQuery}>, reply: FastifyReply) {
     try {
-        const stream = await prisma.stream.findFirst({
+        const streams = await prisma.stream.findMany({
+            take: request.query.limit,
+            skip: request.query.limit * (request.query.page),
             where: {
-                userUserId: request.params.userUserId
+                userUserId: request.user.userId
             }
         })
-        if (!stream) return reply.status(400).send({ error: "User does not exist" });
 
-        return reply.status(200).send(stream);
+        return reply.status(200).send(streams);
     } catch (error) {
         console.log(error);
         return reply.status(500).send(error);
@@ -60,11 +63,11 @@ export async function getStream(request: FastifyRequest<{ Params: GetStreamParam
     }
 }
 
-export async function getStreams(request: FastifyRequest<{ Querystring: GetStreamsQuery }>, reply: FastifyReply) {
+export async function getStreams(request: FastifyRequest<{ Querystring: PaginationQuery }>, reply: FastifyReply) {
     try {
-        const limit = request.query.limit ? request.query.limit : 10;
         const streams = await prisma.stream.findMany({
-            take: limit,
+            take: request.query.limit,
+            skip: request.query.limit * (request.query.page),
             select: {
                 streamId:true,
                 thumbnailUrl:true,
@@ -87,23 +90,23 @@ export async function getStreams(request: FastifyRequest<{ Querystring: GetStrea
 export async function updateStream(request: FastifyRequest<{ Params: {streamId: string },Body: UpdateStreamInput }>, reply: FastifyReply) {
     const input = request.body;
     const streamId = request.params.streamId // Assuming you're passing streamId as a parameter
-
     try {
+        if(!await streamExists(streamId, request.user.userId)) return reply.status(400).send({ error: "stream does not exist" });
         const updatedStream = await prisma.stream.update({
             where: {
-                streamId: streamId
+                streamId: streamId,
+                userUserId: request.user.userId
             },
             data: {
                 thumbnailUrl: input.thumbnailUrl,
                 startTimestamp: new Date(input.startTimestamp),
-                endTimestamp: new Date(input.endTimestamp),
-                storageUrl: input.storageUrl || "",
+                endTimestamp: new Date(input.endTimestamp?input.endTimestamp:input.startTimestamp),
+                storageUrl: input.storageUrl,
                 title: input.title,
                 description: input.description,
-                tags: input.tags || ""
+                tags: input.tags
             }
         });
-
         return reply.status(200).send(updatedStream);
     } catch (error) {
         console.error("Error updating stream:", error);
@@ -113,16 +116,78 @@ export async function updateStream(request: FastifyRequest<{ Params: {streamId: 
 
 export async function deleteStream(request: FastifyRequest<{ Params: { streamId: string } }>, reply: FastifyReply) {
     const streamId = request.params.streamId;
-
     try {
+        if(!await streamExists(streamId, request.user.userId)) return reply.status(400).send({ error: "stream does not exist" });
         const deletedStream = await prisma.stream.delete({
             where: {
-                streamId: streamId
+                streamId: streamId,
+                userUserId: request.user.userId
             }
         });
         return reply.status(200).send(deletedStream);
     } catch (error) {
         console.error("Error deleting stream:", error);
         return reply.status(500).send({ error: "Internal Server Error" });
+    }
+}
+
+
+export async function getMyLikedStreams(request: FastifyRequest<{ Querystring: PaginationQuery }>, reply: FastifyReply) {
+    try {
+        const votes = await prisma.vote.findMany({
+            take: request.query.limit,
+            skip: request.query.limit * (request.query.page),
+            where:{
+                userUserId:request.user.userId,
+                dislike:false
+            },
+            select: {
+                streamStreamId:true
+            }
+        })
+        
+        const streamIds = votes.map(stream => stream.streamStreamId);
+        const streams = await prisma.stream.findMany({
+            where:{
+                streamId:{
+                    in:streamIds
+                }
+            }
+        });
+
+        return reply.status(200).send(streams);
+    } catch (error) {
+        console.log(error);
+        return reply.status(500).send(error);
+    }
+}
+
+export async function getMyDislikedStreams(request: FastifyRequest<{ Querystring: PaginationQuery }>, reply: FastifyReply) {
+    try {
+        const votes = await prisma.vote.findMany({
+            take: request.query.limit,
+            skip: request.query.limit * (request.query.page),
+            where:{
+                userUserId:request.user.userId,
+                dislike:true
+            },
+            select: {
+                streamStreamId:true
+            }
+        })
+        
+        const streamIds = votes.map(stream => stream.streamStreamId);
+        const streams = await prisma.stream.findMany({
+            where:{
+                streamId:{
+                    in:streamIds
+                }
+            }
+        });
+
+        return reply.status(200).send(streams);
+    } catch (error) {
+        console.log(error);
+        return reply.status(500).send(error);
     }
 }
